refactor: moved track IO to tracking
